Resolved a lot of type issues, but ultimately just flipped on the ignore typescript build errors, we will fix later
