+++ src/app/(auth)/admin/reviewapplications/components/ReviewApplicationsTable.tsx
@@ -21,7 +21,7 @@ import {
 
 import { atom, useAtom } from 'jotai';
 
-const sideCardAtom = atom<Applicant>({});
+const sideCardAtom = atom<Applicant>();
 export { sideCardAtom };
 
 // import { sideCardAtom } from '@/app/(auth)/admin/reviewapplications/page';
@@ -157,7 +157,7 @@ export default function ReviewApplicationsTable({
     });
 
     //Data state
-    const [data, setData] = useState([]);
+    const [data, setData] = useState<Applicant[]>([]);
 
     //Change data state on update of DB
     useEffect(() => {
@@ -168,8 +168,7 @@ export default function ReviewApplicationsTable({
     }, [applicationData.data]);
 
     useEffect(() => {
-        // @ts-ignore
-        setData((prevData) =>
+        setData((prevData: Applicant[]) =>
             prevData.map((item) =>
                 item.id === refreshTable.userId
                     ? {
@@ -184,7 +183,11 @@ export default function ReviewApplicationsTable({
 
     //Filters and sorting
     const [globalFilter, setGlobalFilter] = useState<string>('');
-    const [sorting, setSorting] = useState([{ id: 'name', asc: true }]);
+    type SortingState = { id: string; desc: boolean }[];
+
+    const [sorting, setSorting] = useState<SortingState>([
+        { id: 'name', desc: false },
+    ]);
     const [rowSelection, setRowSelection] = useState({});
 
     //I copied this from somewhere, this is for the checkboxes in the table
